+++ src/dd/logon.ts
@@ -394,17 +394,16 @@ module Logon {
                     if (!Ring.have($.player.rings, Ring.theOne) && !$.access.sysop) {
                         $.player.blessed = ''
                         vt.out(vt.yellow, vt.bright, '\nYour shining aura ', vt.normal, 'fades ', vt.faint, 'away.')
-                        PC.activate($.online)
                     }
                 }
                 if ($.player.cursed) {
                     if (!$.player.coward || Ring.have($.player.rings, Ring.theOne) || $.access.sysop) {
                         $.player.cursed = ''
                         vt.out(vt.black, vt.bright, '\nThe dark cloud has been lifted.')
-                        PC.activate($.online)
                     }
                 }
                 $.player.coward = false
+                PC.activate($.online)
             }
 
             if ($.player.level < 50 && 2 * $.player.jw < $.player.jl) {