+++ src/pc.ts
@@ -6,7 +6,7 @@
 import $ = require('./runtime')
 import db = require('./db')
 import { Access, Armor, Magic, Ring, Weapon } from './items'
-import { armor, carry, vt, weapon } from './lib'
+import { armor, carry, log, news, vt, weapon } from './lib'
 import { date2full, dice, fs, int, now, pathTo, romanize, sprintf, titlecase, whole } from './sys'
 
 module pc {
@@ -58,7 +58,7 @@ module pc {
                 let start = 0
                 if (Deed.name[what]) start = Deed.name[what].starting
                 db.run(`INSERT INTO Deeds VALUES ('${pc}', '${what}', ${now().date}, 'Nobody', ${start})`)
-                deed = this.load(pc, what)
+                deed = Deed.load(pc, what)
             }
 
             return deed
@@ -119,7 +119,7 @@ module pc {
         activate(one = $.online, keep = false, confused = false): boolean {
             if (one == $.online) $.online.user = $.player
             one.adept = one.user.wins ? 1 : 0
-            one.pc = this.card(one.user.pc)
+            one.pc = PC.card(one.user.pc)
             one.str = one.user.str
             one.int = one.user.int
             one.dex = one.user.dex
@@ -132,15 +132,15 @@ module pc {
                 one.user.rings.forEach(ring => {
                     rt -= Ring.power(one.user.rings, [ring], 'degrade', 'ability', ability).power * 2
                     rt -= Ring.power(one.user.rings, [ring], 'degrade', 'pc', one.user.pc).power * 3
-                    rt += Ring.power([], [ring], 'upgrade', 'ability', ability).power * this.card(one.user.pc)[a] * 2
-                    rt += Ring.power([], [ring], 'upgrade', 'pc', one.user.pc).power * this.card(one.user.pc)[a] * 3
+                    rt += Ring.power([], [ring], 'upgrade', 'ability', ability).power * PC.card(one.user.pc)[a] * 2
+                    rt += Ring.power([], [ring], 'upgrade', 'pc', one.user.pc).power * PC.card(one.user.pc)[a] * 3
                 })
-                this.adjust(ability, rt, 0, 0, one)
+                PC.adjust(ability, rt, 0, 0, one)
             })
             one.confused = false
             if (confused) return true
 
-            one.who = this.who(one)
+            one.who = PC.who(one)
             one.hp = one.user.hp
             one.sp = one.user.sp
             one.bp = int(one.user.hp / 10)
@@ -161,11 +161,11 @@ module pc {
 
         adjust(ability: ABILITY, rt = 0, pc = 0, max = 0, rpc = $.online) {
             if (max) {
-                rpc.user[`max${ability}`] = this.ability(rpc.user[`max${ability}`], max, 99)
+                rpc.user[`max${ability}`] = PC.ability(rpc.user[`max${ability}`], max, 99)
                 rpc.altered = true
             }
             if (pc) {
-                rpc.user[ability] = this.ability(rpc.user[ability], pc, rpc.user[`max${ability}`])
+                rpc.user[ability] = PC.ability(rpc.user[ability], pc, rpc.user[`max${ability}`])
                 rpc.altered = true
             }
             const a = `to${titlecase(ability)}`
@@ -175,25 +175,100 @@ module pc {
             rpc.user.rings.forEach(ring => {
                 mod -= Ring.power(rpc.user.rings, [ring], 'degrade', 'ability', ability).power * 2
                 mod -= Ring.power(rpc.user.rings, [ring], 'degrade', 'pc', rpc.user.pc).power * 3
-                mod += Ring.power([], [ring], 'upgrade', 'ability', ability).power * this.card(rpc.user.pc)[a] * 2
-                mod += Ring.power([], [ring], 'upgrade', 'pc', rpc.user.pc).power * this.card(rpc.user.pc)[a] * 3
+                mod += Ring.power([], [ring], 'upgrade', 'ability', ability).power * PC.card(rpc.user.pc)[a] * 2
+                mod += Ring.power([], [ring], 'upgrade', 'pc', rpc.user.pc).power * PC.card(rpc.user.pc)[a] * 3
             })
             if (rt > 100) {
                 mod++
                 rt %= 100
             }
-            rpc[ability] = this.ability(rpc[ability], rt, rpc.user[`max${ability}`], mod)
+            rpc[ability] = PC.ability(rpc[ability], rt, rpc.user[`max${ability}`], mod)
         }
 
         card(dd = 'Spirit'): character {
             let rpc = <character>{}
-            for (let type in this.name) {
-                if (this.name[type][dd]) {
-                    rpc = this.name[type][dd]
+            for (let type in PC.name) {
+                if (PC.name[type][dd]) {
+                    rpc = PC.name[type][dd]
                     break
                 }
             }
             return rpc
+        }
+
+        bless(from: string, why: string, who = $.online) {
+            who.altered = true
+            who.user.coward = false
+            if (who.user.cursed) {
+                who.user.cursed = ''
+                if (who == $.online) {
+                    vt.outln('The ', vt.faint, 'dark cloud', vt.normal, ' has left you.', -1000)
+                    news(`\tlifted curse`)
+                }
+            }
+            else {
+                who.user.blessed = from
+                if (who == $.online) {
+                    vt.outln(vt.yellow, 'You feel ', -400,
+                        vt.bright, 'a shining aura', -500,
+                        vt.normal, ' surround you.', -600)
+                    news(`\t${who.user.handle} ${why}`)
+                }
+            }
+            if (who == $.online) {
+                PC.adjust('str', 110)
+                PC.adjust('int', 110)
+                PC.adjust('dex', 110)
+                PC.adjust('cha', 110)
+                vt.sound('shimmer')
+            }
+        }
+
+        curse(from: user, why: string, who = $.online) {
+            who.altered = true
+            who.user.coward = true
+            if (who == $.online) {
+                vt.sound('crack')
+                PC.adjust('str', -10)
+                PC.adjust('int', -10)
+                PC.adjust('dex', -10)
+                PC.adjust('cha', -10)
+                if (who.user.blessed) {
+                    vt.outln(vt.yellow, vt.bright, 'Your shining aura ', -400,
+                        vt.normal, 'fades ', -500,
+                        vt.faint, 'away.', -600)
+                }
+                else {
+                    who.user.cursed = from.id
+                    vt.outln(vt.magenta, vt.bright, from.handle, -400,
+                        vt.normal, ' curses you ', -500,
+                        vt.faint, `${why}!`, -600)
+                    news(`\tcursed by ${from.handle} ${why}`)
+                }
+            }
+            else {
+                if (from.id == $.player.id)
+                    PC.adjust('cha', -1, -1)
+                if (who.user.id == $.king.id) {
+                    from.coward = true
+                    vt.sound('boom')
+                }
+                else
+                    vt.sound('morph')
+                news(`\t${from.handle} cursed ${who.user.handle}`)
+                if (who.user.blessed) {
+                    log(who.user.id, `\n${from.handle} vanquished your blessedness!`)
+                    vt.outln(vt.yellow, vt.bright, who.who.His, 'shining aura', -400,
+                        vt.normal, ' fades ', -500,
+                        vt.faint, 'away.', -600)
+                }
+                else {
+                    log(who.user.id, `\n${$.player.handle} cursed you ${why}!`)
+                    who.user.cursed = $.player.id
+                    vt.outln(vt.faint, 'A dark cloud hovers over ', who.who.him, '.', -1000)
+                }
+            }
+            who.user.blessed = ''
         }
 
         encounter(where = '', lo = 2, hi = 99): active {
@@ -208,7 +283,7 @@ module pc {
             if (rs.length) {
                 let n = dice(rs.length) - 1
                 rpc.user.id = rs[n].id
-                this.load(rpc)
+                PC.load(rpc)
             }
             return rpc
         }
@@ -224,7 +299,7 @@ module pc {
         }
 
         expout(xp: number, awarded = true): string {
-            const gain = int(100 * xp / (this.experience($.player.level) - this.experience($.player.level - 1)))
+            const gain = int(100 * xp / (PC.experience($.player.level) - PC.experience($.player.level - 1)))
             let out = (xp < 1e+8 ? xp.toString() : sprintf('%.4e', xp)) + ' '
             if (awarded && gain && $.online.int >= 90) {
                 out += vt.attr(vt.off, vt.faint, '(', vt.bright
@@ -279,7 +354,7 @@ module pc {
             let user: user = 'user' in rpc ? rpc.user : rpc
             if (user.handle) user.handle = titlecase(user.handle)
             if (db.loadUser(user)) {
-                if ('altered' in rpc) this.activate(rpc)
+                if ('altered' in rpc) PC.activate(rpc)
                 //  restore NPC with static fields
                 if (user.id[0] == '_' && user.id != "_SYS") {
                     let npc = db.fillUser(NPC[user.id], user)
@@ -354,7 +429,7 @@ module pc {
                 fs.accessSync(userPNG, fs.constants.F_OK)
                 userPNG = `user/${rpc.user.id}`
             } catch (e) {
-                userPNG = (this.name['player'][rpc.user.pc] || this.name['immortal'][rpc.user.pc] ? 'player' : 'monster') + '/' + rpc.user.pc.toLowerCase() + (rpc.user.gender == 'F' ? '_f' : '')
+                userPNG = (PC.name['player'][rpc.user.pc] || PC.name['immortal'][rpc.user.pc] ? 'player' : 'monster') + '/' + rpc.user.pc.toLowerCase() + (rpc.user.gender == 'F' ? '_f' : '')
             }
             vt.profile({ png: userPNG, handle: rpc.user.handle, level: rpc.user.level, pc: rpc.user.pc, effect: effect })
             vt.title(`${rpc.user.handle}: level ${rpc.user.level} ${rpc.user.pc} ${meta}`)
@@ -363,19 +438,19 @@ module pc {
         random(type?: string): string {
             let pc: string = ''
             if (type) {
-                let i = dice(Object.keys(this.name[type]).length)
+                let i = dice(Object.keys(PC.name[type]).length)
                 let n = i
-                for (let dd in this.name[type])
+                for (let dd in PC.name[type])
                     if (!--n) {
                         pc = dd
                         break
                     }
             }
             else {
-                let i = dice(this.total - 2)    //  less Spirit and Novice
+                let i = dice(PC.total - 2)    //  less Spirit and Novice
                 let n = i + 2
-                for (type in this.name) {
-                    for (let dd in this.name[type])
+                for (type in PC.name) {
+                    for (let dd in PC.name[type])
                         if (!--n) {
                             pc = dd
                             break
@@ -399,11 +474,11 @@ module pc {
             }
             //  assign & fetch the playing class
             if (pc) user.pc = pc
-            const rpc = this.card(user.pc)
+            const rpc = PC.card(user.pc)
             //  apply character class attributes
             user = db.fillUser(user.pc, user)
             user.level = level
-            if (user.level > 1) user.xp = this.experience(user.level - 1, 1, user.int)
+            if (user.level > 1) user.xp = PC.experience(user.level - 1, 1, user.int)
             if (!user.xplevel || user.xplevel > user.level) user.xplevel = user.level
             if (!user.keyseq) PC.newkeys(user)
             //  level up
@@ -495,8 +570,8 @@ module pc {
                 if ((user.int += rpc.toInt) > user.maxint) user.int = user.maxint
                 if ((user.dex += rpc.toDex) > user.maxdex) user.dex = user.maxdex
                 if ((user.cha += rpc.toCha) > user.maxcha) user.cha = user.maxcha
-                user.hp += this.hp(user)
-                user.sp += this.sp(user)
+                user.hp += PC.hp(user)
+                user.sp += PC.sp(user)
             }
             return user
         }
@@ -569,7 +644,7 @@ module pc {
 
         status(profile: active) {
             vt.action('clear')
-            this.portrait(profile)
+            PC.portrait(profile)
 
             const line = '------------------------------------------------------'
             const space = '                                                      '
@@ -615,10 +690,10 @@ module pc {
             vt.out(sprintf('%-20s', (profile.user.wins ? `${romanize(profile.user.wins)}.` : '')
                 + profile.user.immortal + '.' + profile.user.level + ` (${profile.user.calls})`))
             vt.out(vt.cyan, ' Need: ', vt.white)
-            if (this.experience(profile.user.level, undefined, profile.user.int) < 1e+8)
-                vt.out(sprintf('%-15f', this.experience(profile.user.level, undefined, profile.user.int)))
+            if (PC.experience(profile.user.level, undefined, profile.user.int) < 1e+8)
+                vt.out(sprintf('%-15f', PC.experience(profile.user.level, undefined, profile.user.int)))
             else
-                vt.out(sprintf('%-15.7e', this.experience(profile.user.level, undefined, profile.user.int)))
+                vt.out(sprintf('%-15.7e', PC.experience(profile.user.level, undefined, profile.user.int)))
             vt.outln(' ', vt.reset, vt.blue, vt.faint, '|')
 
             vt.out(vt.blue, vt.faint, '|', vt.Blue, vt.cyan, vt.bright)
@@ -649,7 +724,7 @@ module pc {
 
             if (profile.user.blessed) {
                 let who: user = { id: profile.user.blessed }
-                if (!this.load(who)) {
+                if (!PC.load(who)) {
                     if (profile.user.blessed == 'well')
                         who.handle = 'a wishing well'
                     else
@@ -662,7 +737,7 @@ module pc {
 
             if (profile.user.cursed) {
                 let who: user = { id: profile.user.cursed }
-                if (!this.load(who)) {
+                if (!PC.load(who)) {
                     if (profile.user.cursed == 'wiz!')
                         who.handle = 'a doppelganger!'
                     else
@@ -815,7 +890,7 @@ module pc {
 
             vt.out(vt.blue, vt.faint, '|', vt.Blue, vt.cyan, vt.bright)
             vt.out(' Jousting: ', vt.white)
-            vt.out(sprintf('%-20s', profile.user.jw + ':' + profile.user.jl + ` (${this.jousting(profile)})`))
+            vt.out(sprintf('%-20s', profile.user.jw + ':' + profile.user.jl + ` (${PC.jousting(profile)})`))
             vt.out(vt.cyan, 'Plays: ', vt.white)
             vt.out(sprintf('%-15s', profile.user.plays))
             vt.outln(' ', vt.reset, vt.blue, vt.faint, '|')
@@ -831,17 +906,17 @@ module pc {
 
         wearing(profile: active) {
             if (isNaN(+profile.user.weapon)) {
-                vt.outln('\n', this.who(profile).He, profile.weapon.text, ' ', weapon(profile)
+                vt.outln('\n', PC.who(profile).He, profile.weapon.text, ' ', weapon(profile)
                     , $.from == 'Dungeon' ? -300 : !profile.weapon.shoppe ? -500 : -100)
             }
             if (isNaN(+profile.user.armor)) {
-                vt.outln('\n', this.who(profile).He, profile.armor.text, ' ', armor(profile)
+                vt.outln('\n', PC.who(profile).He, profile.armor.text, ' ', armor(profile)
                     , $.from == 'Dungeon' ? -300 : !profile.armor.armoury ? -500 : -100)
             }
             if (!$.player.novice && $.from !== 'Dungeon' && profile.user.sex == 'I') for (let i in profile.user.rings) {
                 let ring = profile.user.rings[i]
                 if (!+i) vt.outln()
-                vt.out(this.who(profile).He, 'has ', vt.cyan, vt.bright, ring, vt.normal)
+                vt.out(PC.who(profile).He, 'has ', vt.cyan, vt.bright, ring, vt.normal)
                 if ($.player.emulation == 'XT') vt.out(' ', Ring.name[ring].emoji)
                 vt.outln(' powers ', vt.reset, 'that can ', Ring.name[ring].description, -100)
             }