+++ src/types.ts
@@ -5,7 +5,7 @@
 
 type ABILITY = string | 'str' | 'int' | 'dex' | 'cha'
 type DEED = 'big blast' | 'blast' | 'jl' | 'jw' | 'killed' | 'kills' | 'levels' | 'melee' | 'plays' | 'retreats' | 'steals' | 'tl' | 'tw'
-type POWER = string | 'buff' | 'cast' | 'degrade' | 'hp' | 'identify' | 'initiate' | 'joust' | 'melee' | 'resist' | 'resurrect' | 'ring' | 'skip' | 'sp' | 'steal' | 'taxes' | 'teleport' | 'upgrade'
+type POWER = 'buff' | 'cast' | 'curse' | 'degrade' | 'hp' | 'identify' | 'initiate' | 'joust' | 'melee' | 'resist' | 'resurrect' | 'ring' | 'skip' | 'sp' | 'steal' | 'taxes' | 'teleport' | 'upgrade'
 type POWTO = 'ability' | 'magic' | 'melee' | 'pc' | 'spell'
 type GIFT = '' | 'armor' | 'chest' | 'magic' | 'map' | 'poison' | 'potion' | 'ring' | 'weapon' | 'xmagic'
 type MAP = '' | 'map' | 'magic map' | `Marauder's map`